refactor: SameNumberCountInALotto 생성자 변경, LottoNumbers 멤버변수추가
